pyproject.toml
[project]
name = "mcp-git-ingest"
version = "0.1.0"
description = "MCP Git Ingest: GitHub repository analysis tools"
authors = [
    {name = "Adhika Setya Pramudita", email = "adhika.setya.p@gmail.com"},
]
dependencies = [
    "fastmcp",
    "gitpython",
    "uvicorn"
]
requires-python = ">=3.8"
license = {text = "MIT"}

[project.scripts]
mcp-git-ingest = "mcp_git_ingest.main:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_git_ingest"]
